Added schema caching to Schema.cs
How it works:
Schema web connection is opened and headers read.
If cache file is not present or its creation time is older than "Last-Modified" of web schema Schema is pulled from api.steampowered.com, put into cache file and  cache file creation time is set equal to Schema Last-modified.
If cache file is present and is up-to-date the Schema is downloaded from the file.