Fix project build order in VS and MD.
Proto-buf MUST be built before Steamkit.

For MonoDevelop (Xamarin Studio) the order of projects in the solution seems important. MD has no GUI for editing build order so this seems to be the only way to specify it.

The MD changes do not break VS2010 support. But the visual studio GUI for build orders was used to change the build order. (VS wins this round).