Replaces Current Trade Event Checking Fixes #219
This changes the current way of checking events processed. Instead of
having a number, and using a for loop, this adds the events that are
processed into a list. Every time a status is obtained, it checks every
event to make sure it hasn't missed anything. I did have to add a method
that checks if the list contains the event, as List.Contains() was not
working, and kept triggering the events multiple times.

Ever since I have put this in, the bots have not missed any removing of
items, and it has been well over a week, with a decent amount of trades.