Adds "exec" command and other BotManager tweaks.
Allows commands to be given directly to bots via the BotManager
console.This command is handled in the UserHandler's `OnBotCommand`.

Other changes include threading the BotManager commands and
slightlychanging the way accounts are authorized while
`UseSeparateProcesses  = true`.

Extended functionality of the "start", "stop", and "auth" commands to
allow usernames as well as indexes.

Also fixed the `Console.Title` being set to the last bot to be
configured while in Separate Processes mode.