Remove submodule note and other volatile info from README
In addition to removing the stuff pertaining to downloading submodules I also removed the note about not downloading the zip. I've tested building via the snapshot myself and it works fine. Finally, I removed some information about compiling the bot so that it won't have to be changed as often in the future.