Bot/log disposal fix
- Previously, if the bot was forcibly, or even due to an error stopped
and then started again without restarting the entire application, the
program would throw an error while trying to write to the log file,
because it was still open.
The function for the disposal was already properly defined, but never
called if the GC never picked it up (which it did not on my machine).